Customers shown a warning when booking in the next 12 hours

+++ src/main/java/com/LilliputSalon/SalonApp/service/AppointmentManagerService.java
@@ -382,7 +382,7 @@ public class AppointmentManagerService {
 
 		    Availability availability =
 		        availabilityRepo.findByUser_IdAndWorkDate(
-		            stylistId.longValue(), date
+		            stylistId, date
 		        );
 
 		    if (availability == null) {
